refactor: move IInputOutput and ConsoleIO to appropriate architecture layers
- Moved IInputOutput interface to MooGame.App.Interfaces
- Moved ConsoleIO implementation to MooGame.App.View
- Updated namespaces and using directives accordingly
- No logic changes made; preserves existing behavior
- Improves project structure by clearly separating interface and console-specific logic
