+++ MooGame.App/View/ConsoleIO.cs
@@ -1,6 +1,7 @@
-﻿using MooGame.App;
+﻿using MooGame.App.Helper;
+using MooGame.App.Interfaces;
 
-namespace MooGame.App;
+namespace MooGame.App.View;
 
 public class ConsoleIO : IInputOutput
 {
